Update README.md
Coming to terms- The goal was to write a paper that explains how to do something clearly . I composed a document on how to start a successful business. 

Understanding what you have/ Forwarding - I believe that I accomplished this goal could have done so in a clearer manner

Countering- I believe I could have given a better idea of who the audience is. Could simplify some topics and expand others

Looking ahead- If I ever compose another how to document I will choose a task that is not as broad as starting a successful business and will attempt to communicate my thoughts and ideas in a clearer manner

@matthewjmiller